Added SettingsScene, Configurable nicknames and scoreboard file and functionality to all buttons.
Added Settings button with a Settings Scene that containes 2 fields where the player can set his
nickname and the file where to retain his score and a return to menu button.
Player names are functional and are written correctly in the specified file.
If someone doesn't fill thoes fields, the default player name is <<Unkonwn>> and the score file is <<Scores.txt>>
All buttons added until now work as intended.

Signed-off-by: alin-ghita <alin_g2008@yahoo.com>